{IB-21} Refactoring kodu oraz możliwość dodania nowego pracownika kościelnego.
