+++ src/main/java/com/zakrzewski/intentionbook/IntentionBookApplication.java
@@ -1,24 +1,9 @@
 package com.zakrzewski.intentionbook;
 
-import com.zakrzewski.intentionbook.abstractClass.ChurchWorker;
-import com.zakrzewski.intentionbook.entities.BookOfIntentionModel;
-import com.zakrzewski.intentionbook.entities.PriestModel;
-import com.zakrzewski.intentionbook.entities.SacristanModel;
-import com.zakrzewski.intentionbook.enums.AccessEnum;
-import com.zakrzewski.intentionbook.repositories.BookOfIntentionRepository;
-import com.zakrzewski.intentionbook.repositories.ChurchWorkerRepository;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.context.event.ApplicationReadyEvent;
 import org.springframework.context.annotation.Bean;
-import org.springframework.context.event.EventListener;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.web.client.RestTemplate;
-
-import java.time.LocalDate;
-import java.time.LocalTime;
 
 
 @SpringBootApplication
