+++ src/main/resources/application.properties
@@ -1,38 +1,11 @@
 #----------------------------------
 #               PostgreSQL
 #----------------------------------
-#spring.datasource.platform=postgres
-#spring.datasource.url=jdbc:postgresql://localhost:5432/intention_book
-#spring.datasource.username=postgres
-#spring.datasource.password=kanapka
-#spring.jpa.properties.hibernate.dialect = org.hibernate.dialect.PostgreSQLDialect
-
-
-spring.datasource.url=xxx
-spring.datasource.username=xx
-spring.datasource.password=xxxx
+spring.datasource.platform=postgres
+spring.datasource.url=jdbc:postgresql://localhost:5432/intention_book
+spring.datasource.username=postgres
+spring.datasource.password=kanapka
 spring.jpa.properties.hibernate.dialect = org.hibernate.dialect.PostgreSQLDialect
-
-#----------------------------------
-#               RemoteMySql
-#----------------------------------
-
-#spring.datasource.url=jdbc:mysql://remotemysql.com:3306/Z9kgCAHj6U
-#spring.datasource.username=Z9kgCAHj6U
-#spring.datasource.password=jEBvKKRY09
-#spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
-#spring.jpa.properties.hibernate.dialect = org.hibernate.dialect.MySQL5Dialect
-
-
-#----------------------------------
-#               H2Database
-#----------------------------------
-
-#spring.datasource.url=jdbc:h2:mem:db_intention
-#spring.h2.console.enabled=true
-#spring.datasource.driverClassName=org.h2.Driver
-#spring.datasource.username=sa
-#spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
 
 #----------------------------------
 #               Hibernate
@@ -40,4 +13,4 @@ spring.jpa.properties.hibernate.dialect = org.hibernate.dialect.PostgreSQLDialec
 spring.jpa.show-sql=true
 spring.jpa.properties.hibernate.temp.use_jdbc_metadata_defaults=false
 spring.jpa.generate-ddl=true
-spring.jpa.hibernate.ddl-auto=create
+spring.jpa.hibernate.ddl-auto=update
