+++ src/main/resources/templates/create-new-worker.html
@@ -1,0 +1,54 @@
+<!DOCTYPE html>
+<html lang="PL" xmlns:th="http://www.thymeleaf.org" xmlns="http://www.w3.org/1999/xhtml">
+<head>
+    <meta charset="utf-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
+    <script type="text/javascript" src="https://code.jquery.com/jquery-3.3.1.min.js"></script>
+    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/css/bootstrap.min.css" integrity="sha384-Gn5384xqQ1aoWXA+058RXPxPg6fy4IWvTNh0E263XmFcJlSAwiGgFAW/dAiS6JXm" crossorigin="anonymous"/>
+    <script src="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/js/bootstrap.min.js" integrity="sha384-JZR6Spejh4U02d8jOt6vLEHfe/JQGiRRSQQxSfFWpi1MquVdAyjUar5+76PVCmYl" crossorigin="anonymous"></script>
+    <title>Utwórz nowego pracownika</title>
+</head>
+<body>
+<nav class="navbar navbar-expand-sm bg-dark navbar-dark">
+    <ul class="navbar-nav">
+        <li class="nav-item">
+            <a class="nav-link" th:href="@{/}">Pokaż intencję</a>
+        </li>
+        <li class="nav-item">
+            <a class="nav-link" th:href="@{show-all-workers}">Pokaż pracowników</a>
+        </li>
+    </ul>
+</nav>
+<div class="formSaveWorker">
+    <form th:action="@{/add-new-worker}" method="post" th:object="${newWorker}">
+        <div class="form-group col-lg-3">
+            <label>Login:</label>
+            <input type="text" class="form-control" th:field="*{workerLogin}" th:value="*{workerLogin}" id="workerLogin" placeholder="Wpisz login" th:required="required">
+        </div>
+        <div class="form-group col-lg-3">
+            <label>Hasło:</label>
+            <input type="password" class="form-control" th:field="*{workerPassword}" th:value="*{workerPassword}" id="workerPassword" placeholder="Wpisz hasło" th:required="required">
+        </div>
+        <div class="form-group col-lg-3">
+            <label>Imię:</label>
+            <input type="text" class="form-control" th:field="*{firstName}" th:value="*{firstName}" id="firstName" placeholder="Wpisz imię" th:required="required">
+        </div>
+        <div class="form-group col-lg-3">
+            <label>Nazwisko:</label>
+            <input type="text" class="form-control" th:field="*{lastName}" th:value="*{lastName}" id="lastName" placeholder="Wpisz nazwisko" th:required="required">
+        </div>
+        <div class="form-group col-lg-3">
+            <label>Rola:</label>
+            <select class="form-control" th:field="*{accessEnum}"  id="accessEnum" name="accessEnum" th:required="required">
+                <option th:each="accessEnum : ${T(com.zakrzewski.intentionbook.enums.AccessEnum).values()}"
+                        th:value="${accessEnum.getAuthority()}"
+                        th:text="${accessEnum.getAuthority()}"></option>
+            </select>
+        </div>
+        <div class="form-group col-lg-3">
+            <button type="submit" class="btn btn-primary">Dodaj pracownika</button>
+        </div>
+    </form>
+</div>
+</body>
+</html>
