+++ src/main/resources/templates/index.html
@@ -21,6 +21,9 @@
             <a sec:authorize="isAuthenticated()" class="nav-link" th:href="@{show-all-workers}">Pokaż pracowników</a>
         </li>
         <li class="nav-item">
+            <a sec:authorize="isAuthenticated()" class="nav-link" th:href="@{create-new-worker-page}">Dodaj pracownika</a>
+        </li>
+        <li class="nav-item">
             <a sec:authorize="isAuthenticated()" class="nav-link" th:href="@{calendar-info}">Kalendarz liturgiczny</a>
         </li>
         <li class="nav-item">
